connecting articles to api
